+++ server/routes/summary_routes.ts
@@ -8,6 +8,7 @@ import { IRouter, OpenSearchClient } from '../../../../src/core/server';
 import { SUMMARY_ASSISTANT_API } from '../../common/constants/llm';
 import { getOpenSearchClientTransport } from '../utils/get_opensearch_client_transport';
 import { searchAgent } from './get_agent';
+import { detectIndexType } from './index_type_detect';
 import { AssistantServiceSetup } from '../services/assistant_service';
 import { handleError } from './error_handler';
 import { AgentNotFoundError } from './errors';
@@ -50,16 +51,25 @@ export function registerSummaryAssistantRoutes(
       },
     },
     router.handleLegacyErrors(async (context, req, res) => {
-      const client = await getOpenSearchClientTransport({
-        context,
-        dataSourceId: req.query.dataSourceId,
-      });
-      const assistantClient = assistantService.getScopedClient(req, context);
-      const agentConfigId =
-        req.body.index && req.body.dsl
-          ? LOG_PATTERN_SUMMARY_AGENT_CONFIG_ID
-          : SUMMARY_AGENT_CONFIG_ID;
       try {
+        const client = await getOpenSearchClientTransport({
+          context,
+          dataSourceId: req.query.dataSourceId,
+        });
+        const assistantClient = assistantService.getScopedClient(req, context);
+        let isLogIndex = false;
+        if (req.body.index) {
+          isLogIndex = await detectIndexType(
+            client,
+            assistantClient,
+            req.body.index,
+            req.query.dataSourceId
+          );
+        }
+        const agentConfigId =
+          req.body.index && req.body.dsl && isLogIndex
+            ? LOG_PATTERN_SUMMARY_AGENT_CONFIG_ID
+            : SUMMARY_AGENT_CONFIG_ID;
         const response = await assistantClient.executeAgentByConfigName(agentConfigId, {
           context: req.body.context,
           question: req.body.question,