+++ CHANGELOG.md
@@ -11,17 +11,18 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - add data to summary response post processing ([#436](https://github.com/opensearch-project/dashboards-assistant/pull/436))
 - add flag to control if display conversation list ([#438](https://github.com/opensearch-project/dashboards-assistant/pull/438))
 - when open chatbot, load the last conversation automatically ([#439](https://github.com/opensearch-project/dashboards-assistant/pull/439))
+- add index type detection ([#454](https://github.com/opensearch-project/dashboards-assistant/pull/454))
 
 ### Enhancements
 
 - remove os_insight agent ([#452](https://github.com/opensearch-project/dashboards-assistant/pull/452))
+- Hide the assistant entry when there isn't data2summary agent ([#417](https://github.com/opensearch-project/dashboards-assistant/pull/417))
 
 ### Bug Fixes
 
 - fixed incorrect message id field used ([#378](https://github.com/opensearch-project/dashboards-assistant/pull/378))
 - Improve alert summary with backend log pattern experience ([#389](https://github.com/opensearch-project/dashboards-assistant/pull/389))
-- fixed in context feature returning 500 error if workspace is invalid to returning 4XX [#429](https://github.com/opensearch-project/dashboards-assistant/pull/429)
-
+- fixed in context feature returning 500 error if workspace is invalid to returning 4XX ([#429](https://github.com/opensearch-project/dashboards-assistant/pull/429))([#458](https://github.com/opensearch-project/dashboards-assistant/pull/458))
 
 ### Infrastructure